Made graphs more organized
