+++ cair_app_v2/lib/models/graphs.dart
@@ -1,17 +1,14 @@
 import 'package:flutter/material.dart';
+import './../util/dataliststream.dart';
 
 class Graph {
-  String id;
   String title;
-  double amount;
   DateTime date;
-  List<double> data;
+  int column;
 
   Graph({
-    @required this.id,
     @required this.title,
-    @required this.amount,
     @required this.date,
-    @required this.data,
+    @required this.column,
   });
 }
